[Refactor] users, user 통합, feed resolver 일부 리펙토링
- auth, user로 분리
- dbparser 최근 함수로 변경
- user에서 pubsub 파일 분리. key도 분리
