+++ server/src/api/user/user.resolvers.ts
@@ -1,66 +1,53 @@
-import uploadToObjStorage from '../../middleware/uploadToObjStorage';
-import {
-  MutationSignUpArgs,
-  User,
-  MutationResolvers,
-  QueryResolvers,
-  QueryGetUserArgs
-} from '../../types';
+import { withFilter } from 'graphql-subscriptions';
+import { LOGIN_CHANNEL, LOGOUT_CHANNEL } from './channels';
 import { requestDB } from '../../utils/requestDB';
-import { parseNodeResult } from '../../utils/parseDB';
-import { encodeJWT } from '../../utils/jwt';
-import SameEmailError from '../../errors/EmailAlreadyExistsError';
-import { FIND_USER_WITH_EMAIL_QUERY } from '../../schema/user/query';
-import isAuthenticated from '../../utils/isAuthenticated';
+import {
+  FIND_USER_WITH_EMAIL_QUERY,
+  FIND_FRIENDS_QUERY,
+  CHECK_FRIEND_QUERY,
+  FIND_RELATIONSHIP_WITH_USER
+} from '../../schema/user/query';
+import { getNode, getFirstKeyValue } from '../../utils/parseDB';
+import { socketCountWithEmail } from '../../utils/socketManager';
 import { findUserWithEmail } from '../../schema/user/user';
-import createDBError from '../../errors/createDBError';
+import { QueryResolvers, QueryGetUserArgs, User } from '../../types';
+import isAuthenticated from '../../utils/isAuthenticated';
 import { parseResultRecords } from '../../utils/parseData';
-
-const checkIsEmailExist = async (email): Promise<void> => {
-  const sameUsers = await requestDB(FIND_USER_WITH_EMAIL_QUERY, { email });
-
-  if (parseNodeResult(sameUsers).length) {
-    throw new SameEmailError();
-  }
-};
-
-const getFileUrl = async file => {
-  const { filename, createReadStream } = await file;
-
-  const res = await uploadToObjStorage(createReadStream(), filename);
-  return res.Location;
-};
-
-const getUrlWhenFileExists = file => {
-  if (file) {
-    return getFileUrl(file);
-  }
-  return null;
-};
-
-const createUser = async info => {
-  const result = await requestDB(
-    `CREATE (u:User {nickname: $nickname, hometown: $hometown, residence: $residence, email: $email ${
-      info.thumbnail ? ', thumbnail: $thumbnail' : ''
-    }
-      }) RETURN u`,
-    info
-  );
-  return result[0].get(0).properties;
-};
-
-const Mutation: MutationResolvers = {
-  signUp: async (_, args: MutationSignUpArgs, { res }): Promise<User> => {
-    await checkIsEmailExist(args.email);
-    const thumbnail = await getUrlWhenFileExists(args.file);
-    const user = await createUser({ ...args, thumbnail });
-    const token: string = encodeJWT({ email: args.email });
-    res.cookie('token', token);
-    return user;
-  }
-};
+import createDBError from '../../errors/createDBError';
 
 const Query: QueryResolvers = {
+  friends: async (_, __, { req }): Promise<any[]> => {
+    const result = await requestDB(FIND_FRIENDS_QUERY, {
+      email: req.email
+    });
+    const friends = await parseResultRecords(result);
+
+    return friends.map(({ friends: user }) => {
+      const { email } = user;
+
+      user.status = socketCountWithEmail.has(email) ? 'online' : 'offline';
+      return user;
+    });
+  },
+  userWithEmail: async (_, { email }) => {
+    const result = await requestDB(FIND_USER_WITH_EMAIL_QUERY, { email });
+    const user = await getNode(result);
+    return user;
+  },
+  friendsWithUserEmail: async (_, { email }): Promise<User[]> => {
+    const result = await requestDB(FIND_FRIENDS_QUERY, { email });
+    const friends = parseResultRecords(result);
+
+    return friends.map(friend => friend.friends);
+  },
+  findRelation: async (_, { email: userEmail }, { req }) => {
+    const result = await requestDB(FIND_RELATIONSHIP_WITH_USER, {
+      myEmail: req.email,
+      userEmail
+    });
+    const relation = await getFirstKeyValue(result);
+    return relation;
+  },
   me: async (_, __, { req }): Promise<User> => {
     isAuthenticated(req);
     const { email } = req;
@@ -88,6 +75,24 @@ const Query: QueryResolvers = {
 };
 
 export default {
-  Mutation,
-  Query
+  Query,
+  Subscription: {
+    updateUserState: {
+      resolve: payload => payload,
+      subscribe: withFilter(
+        (_, __, { pubsub }) => {
+          return pubsub.asyncIterator([LOGIN_CHANNEL, LOGOUT_CHANNEL]);
+        },
+        async (payload, _, { email }) => {
+          
+          const result = await requestDB(CHECK_FRIEND_QUERY, {
+            email: payload.email,
+            friendEmail: email
+          });
+          const friend = await getNode(result);
+          return !!friend;
+        }
+      )
+    }
+  }
 };
