+++ server/src/utils/parseDB.ts
@@ -1,9 +1,5 @@
-interface IKey<T> {
-  [key: string]: T;
-}
-
 const parseNodeResult = result => {
-  const returnArr: Array<IKey<string | number>> = [];
+  const returnArr: any = [];
   for (const item of result) {
     returnArr.push(item.get(0).properties);
   }
