+++ server/src/utils/pubsub.ts
@@ -1,7 +1,6 @@
 import { RedisPubSub } from 'graphql-redis-subscriptions';
-import { LOGIN_CHANNEL, LOGOUT_CHANNEL } from './channels';
 
-export const pubsub = new RedisPubSub({
+export default new RedisPubSub({
   connection: {
     host: '127.0.0.1',
     port: 6379,
@@ -11,11 +10,3 @@ export const pubsub = new RedisPubSub({
     }
   }
 });
-
-export const loginPublish = args => {
-  pubsub.publish(LOGIN_CHANNEL, args);
-};
-
-export const logoutPublish = args => {
-  pubsub.publish(LOGOUT_CHANNEL, args);
-};
